WPB-23631: Move `Spar.Sem.VerdictFormatStore` in `Wire.VerdictFormatStore`
Move the VerdictFormatStore effect + Cassandra/Mem backends to
Wire.VerdictFormatStore(.Cassandra/.Mem). Fold the VerdictFormatCon/Row row
codec and from/toVerdictFormat out of Spar.Data.Instances into
Wire.VerdictFormatStore.Cassandra (VerdictFormat/AReqId already in wire-api).
Drop the now-unused nominalDiffToSeconds export from Spar.Data (the .Cassandra
backend duplicates it, since wire-subsystems cannot depend on spar). Update
weeder.toml root. Rewire consumers (CanonicalInterpreter, Spar.API/App,
DataSpec, IdPSpec); delete old modules; register in cabal.
