+++ changelog.d/5-internal/WPB-23631-5
@@ -1,0 +1,1 @@
+Move `Spar.Sem.VerdictFormatStore` in `Wire.VerdictFormatStore`.
