+++ libs/wire-subsystems/src/Wire/VerdictFormatStore.hs
@@ -17,7 +17,7 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Spar.Sem.VerdictFormatStore
+module Wire.VerdictFormatStore
   ( VerdictFormatStore (..),
     store,
     get,
